multi-doc-rag-system/requirements.txt
langchain_community
langchain
langchain_core
python-dotenv
streamlit
langchain-classic
langserve
fastapi
uvicorn
sse_starlette
pypdf
docx2txt
langchain-ollama
faiss-cpu
unstructured
markdown
tiktoken
from_root
sentence-transformers
python-multipart
-e .